add snapshot for app tests
